Added Bosphorous
Updated seed values through Bosphorous
Fixed the "level With 1+ Hangs" metric, was previously not handling red matches properly. Also renamed to "level With Any Hangs"
Replaced "placeholder 1" with a "level And Hanging Rate", which is found by counting the matches in which there is both a level rung and the given team is hanging.
I don't know if/when events will start up again, so I've changed the "outdated version" warning to reflect this.